+++ src/main/java/com/fithub/fithubbackend/domain/user/api/MapController.java
@@ -1,8 +1,7 @@
 package com.fithub.fithubbackend.domain.user.api;
 
 import com.fithub.fithubbackend.domain.user.application.MapService;
-import com.fithub.fithubbackend.domain.user.dto.constants.MapDocumentDto;
-import com.fithub.fithubbackend.domain.user.dto.constants.MapDto;
+import com.fithub.fithubbackend.domain.user.dto.MapDto;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import lombok.RequiredArgsConstructor;
@@ -11,7 +10,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
