+++ src/main/java/com/fithub/fithubbackend/domain/user/application/MapService.java
@@ -1,8 +1,6 @@
 package com.fithub.fithubbackend.domain.user.application;
 
-import com.fithub.fithubbackend.domain.user.dto.constants.MapDto;
-
-import java.util.List;
+import com.fithub.fithubbackend.domain.user.dto.MapDto;
 
 public interface MapService {
     MapDto getLocationByFitness(int page,double x, double y);
